bug: don't need to reload when updating a product
