+++ server/api/index.js
@@ -5,7 +5,6 @@ router.use('/users', require('./users'))
 router.use('/cart', require('./cart'))
 router.use('/products', require('./products'))
 
-
 router.use((req, res, next) => {
   const error = new Error('Not Found')
   error.status = 404
